login issue handled

+++ src/components/MovieTrailer.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useCallback, useEffect, useState } from "react";
 import { AiOutlinePlus, AiOutlinePlayCircle } from "react-icons/Ai";
 import { TiTickOutline } from "react-icons/Ti";
 import { BsFillPlayCircleFill } from "react-icons/Bs";
@@ -18,33 +18,40 @@ const MovieTrailer = ({ videoUrl, img, id, myList }: MovieTrailerProp) => {
   const videoRef = useRef<HTMLVideoElement>(null);
   const [hover, setHover] = useState<boolean>(false);
   const [added, setAdded] = useState<boolean>(false);
-  const {mutate}=useFavoriteMovies()
+  const { mutate } = useFavoriteMovies();
 
-  async function addHandler(id: string) {
+  useEffect(() => {
+    async function fetcher() {
+      if (added) {
+        try {
+          await axios.post("http://localhost:3000/api/modifyFavorites", {
+            movieId: id,
+          });
+          mutate();
+        } catch (error) {
+          console.log(error);
+        }
+      } else {
+        try {
+          await axios.delete(
+            `http://localhost:3000/api/modifyFavorites?movieId=${id}`
+          );
+          mutate();
+        } catch (error) {
+          console.log(error);
+        }
+      }
+    }
+    fetcher();
+  }, [added]);
+  console.log(added);
+  const addHandler = useCallback(async function addHandler(id: string) {
     setAdded(true);
-    if (myList) {
+    if (added) {
       setAdded(false);
     }
-    try {
-      await axios.post("http://localhost:3000/api/modifyFavorites", {
-        movieId: id,
-      });
-      mutate()
+  }, []);
 
-    } catch (error) {
-      console.log(error);
-    }
-  }
-  async function removeHandler(id: string) {
-    try {
-      await axios.delete(
-        `http://localhost:3000/api/modifyFavorites?movieId=${id}`
-      );
-      mutate()
-    } catch (error) {
-      console.log(error);
-    }
-  }
   if (hover) {
     videoRef.current?.play();
   }
@@ -70,8 +77,12 @@ const MovieTrailer = ({ videoUrl, img, id, myList }: MovieTrailerProp) => {
           <button className={styles.btn}>
             <BsFillPlayCircleFill size={40} />
           </button>
-          <button className={styles.btn} onClick={() => removeHandler(id)}>
-            <TiTickOutline style={{ fontSize: "1.2rem" }} />
+          <button className={styles.btn} onClick={() => addHandler(id)}>
+            {added ? (
+              <TiTickOutline style={{ fontSize: "1.2rem" }} />
+            ) : (
+              <AiOutlinePlus style={{ fontSize: "1.2rem" }} />
+            )}
           </button>
         </div>
       </div>
